refactor: better naming for system folder files
