+++ src/index.ts
@@ -1,6 +1,6 @@
 import ExtendedClient from './client/ExtendedClient.js';
-import { processEnv } from './systems/config.js';
-import { connectDB } from './systems/mongo.js';
+import { processEnv } from './systems/configSystem.js';
+import { connectDB } from './systems/mongoSystem.js';
 
 processEnv();
 await connectDB();
